feat: refactor namings
