+++ src/frontend/models/Instruments/Instrument.ts
