+++ src/frontend/models/Instruments/TonejsSynth.ts
