+++ src/frontend/models/Sounds/TonejsSynthSounds.ts
@@ -1,5 +1,5 @@
 import {Sounds} from "./Sounds";
-import {TONEJS_SYNTH_SOUNDS} from "../../vars/TONEJS_SYNTH_SOUNDS";
+import {TONEJS_SYNTH_SOUNDS} from "../../vars/sounds/TONEJS_SYNTH_SOUNDS";
 
 export class TonejsSynthSounds extends Sounds {
     constructor() {
