--- src/frontend/vars/SOUNDS.ts
@@ -1,4 +1,0 @@
-import {SoundObj} from "../models/SoundObj";
-import {TONEJS_SYNTH_SOUNDS} from "./TONEJS_SYNTH_SOUNDS";
-
-export const SOUNDS: SoundObj[] = TONEJS_SYNTH_SOUNDS;
