+++ src/frontend/vars/sounds/DEFAULT_SOUNDS.ts
@@ -1,0 +1,4 @@
+import {SoundObj} from "../../models/SoundObj";
+import {TONEJS_SYNTH_SOUNDS} from "./TONEJS_SYNTH_SOUNDS";
+
+export const DEFAULT_SOUNDS: SoundObj[] = TONEJS_SYNTH_SOUNDS;
